拓扑图修改完善

+++ mis-power/src/main/java/com/wt/mis/dev/entity/BranchBox.java
@@ -33,6 +33,6 @@ public class BranchBox extends BaseEntity {
     * 运维班组编号
     */
     @Column(columnDefinition = " int(11) COMMENT '运维班组编号'")
-    private Integer operationsTeam;
+    private Long operationsTeam;
 
 }
