+++ mis-power/src/main/java/com/wt/mis/dev/entity/Meter.java
@@ -63,7 +63,7 @@ public class Meter extends BaseEntity {
     * 运维班组
     */
     @Column(columnDefinition = " int(11) COMMENT '运维班组'")
-    private Integer operationsTeam;
+    private Long operationsTeam;
 
     /**
      * 单/三相:单相：1 三相 2
