+++ mis-power/src/main/java/com/wt/mis/dev/entity/TransForm.java
@@ -63,6 +63,6 @@ public class TransForm extends BaseEntity {
     * 运维班组
     */
     @Column(columnDefinition = " int(11) COMMENT '运维班组编号'")
-    private Integer operationsTeam;
+    private Long operationsTeam;
 
 }
